Web_App/backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
psycopg2-binary==2.9.9
python-jose[cryptography]==3.3.0
bcrypt==4.1.3
python-dotenv==1.0.1
pydantic[email]==2.7.1
pydantic-settings==2.3.4
python-multipart==0.0.9
